teacher

+++ delete_teacher.php
@@ -2,10 +2,11 @@
 session_start();
 include("db.php");
 
-if(!isset($_SESSION['admin'])){
+if (!isset($_SESSION['role']) || $_SESSION['role'] != "admin") {
     header("Location: index.php");
-    exit;
+    exit();
 }
+
 
 $id = $_GET['id'];
 mysqli_query($conn, "DELETE FROM teachers WHERE id='$id'");
